+++ demo/pendulum_control/main_parallel.py
@@ -3,331 +3,11 @@ import numpy as np
 from autograd import jacobian
 
 import brica
-from brica import Component, VirtualTimeScheduler, Timing
+from brica import VirtualTimeScheduler, Timing
 
-from matchernet import MPCEnv
-from matchernet.state import StateMuSigma
-from matchernet.ekf import MatcherEKF
+from matchernet import MPCEnvBundle, MatcherController, Plan, MatcherILQR, BundlePlan, BundleEKFWithController, MatcherEKF, MPCEnv, MovieWriter, AnimGIFWriter
+
 from pendulum import PendulumDynamics, PendulumCost, PendulumRenderer
-from matchernet import iLQG
-from matchernet import MovieWriter, AnimGIFWriter
-
-from matchernet import Bundle
-from matchernet import utils
-
-
-class BundleEKFWithController(Bundle):
-    def __init__(self, name, dt, f, Q, mu, Sigma, control_src_name):
-        super(BundleEKFWithController, self).__init__(name)
-        self.f = f
-        self._initialize_control_params(dt)
-        self._initialize_state(mu, Sigma)
-        
-        self.Q = Q
-        self.control_src_name = control_src_name
-
-        self.update_component()
-
-    def __call__(self, inputs):
-        for key in inputs:  # key is one of the matcher names
-            if inputs[key] is not None:
-                if key != self.control_src_name:
-                    self.accept_feedback(inputs[key])
-
-        for key in inputs:  # key is one of the matcher names
-            if inputs[key] is not None:
-                if key == self.control_src_name:
-                    u = inputs[key]["u"]
-                    self.step_dynamics(u, self.dt)
-
-        self._countup()
-
-        results = {
-            "mu": self.state.data["mu"],
-            "Sigma": self.state.data["Sigma"],
-            "time_stamp": self.time_stamp,
-            "time_id": self.time_id
-        }
-        return {"state": results}
-
-    def _initialize_control_params(self, dt):
-        self.dt = dt
-        self.time_stamp = 0.0
-        self.time_id = 0
-
-    def _countup(self):
-        self.time_stamp += self.dt
-        self.time_id += 1
-
-    def _initialize_state(self, mu, Sigma):
-        self.state = StateMuSigma(mu, Sigma)
-
-    def accept_feedback(self, fbst):
-        dmu = fbst["mu"]
-        dSigma = fbst["Sigma"]
-        mu = self.state.data["mu"]
-        Sigma = self.state.data["Sigma"]
-
-        self.state.data["mu"] = (mu + dmu).astype(np.float32)
-        self.state.data["Sigma"] = (Sigma + dSigma).astype(np.float32)
-
-    def step_dynamics(self, u, dt):
-        mu = self.state.data["mu"]
-        Sigma = self.state.data["Sigma"]
-        Q = self.Q
-        A = self.f.x(mu, u)
-        F = utils.calc_matrix_F(A, dt)
-        mu = mu + self.f.value(mu, u) * dt
-        Sigma = dt * Q + F.T @ Sigma @ F
-        Sigma = utils.regularize_cov_matrix(Sigma)
-
-        self.state.data["mu"] = mu
-        self.state.data["Sigma"] = Sigma
-        # ["time_stamp"] is updated in the method self._countup()
-
-
-class MPCEnvBundle(Bundle):
-    """
-    MPCEnv bundle class that communicates with matcher.
-    """
-
-    def __init__(self, name, env, R, control_src_name, debug_recorder=None):
-        """
-        Arguments:
-          env
-            MPCEnv instance
-          R
-            Observation noise
-          control_src_name
-            Name of controller src node (Controller Matcher)
-        """
-        super(MPCEnvBundle, self).__init__(name)
-        self.env = env
-        self.R = R
-        self.time_stamp = 0.0
-        self.time_id = 0
-        self.control_src_name = control_src_name
-        self.debug_recorder = debug_recorder
-        
-        self.update_component()
-
-    def __call__(self, inputs):
-        u_dim = self.env.dynamics.u_dim
-
-        # Receive action from Controller Matcher
-        if self.control_src_name in inputs.keys() and inputs[self.control_src_name] is not None:
-            u = inputs[self.control_src_name]["u"]
-        else:
-            u = np.zeros((u_dim,), dtype=np.float32)
-
-        # Step environment with received action
-        # Ignoring env rewards
-        x, _ = self.env.step(u)
-        
-        results = {}
-        results["mu"] = x
-        results["Sigma"] = self.R
-        results["time_stamp"] = self.time_stamp
-        results["time_id"] = self.time_id
-
-        if self.debug_recorder is not None:
-            self.debug_recorder.record(x, u)
-
-        self.time_stamp += self.env.dt
-        self.time_id += 1
-
-        # Send state to matcher
-        return {"state": results}
-
-
-class MatcherController(object):
-    def __init__(self, name, mpcenv_bundle, ekf_bundle, plan_bundle):
-        self.name = name
-        self.results = {}
-
-        self.mpcenv_bundle = mpcenv_bundle
-        self.ekf_bundle = ekf_bundle
-        self.plan_bundle = plan_bundle
-        
-        self.results[self.mpcenv_bundle.name] = {}
-        self.results[self.ekf_bundle.name] = {}
-        # Note: there is no results entry for plan_bundle
-
-        self.update_component()
-
-    def update_component(self):
-        component = Component(self)
-
-        # MatcherController -> MPCEnvBundle
-        component.make_out_port(self.mpcenv_bundle.name)
-        self.mpcenv_bundle.component.make_in_port(self.name)  # -> send u to MPCEnv
-        brica.connect(component, self.mpcenv_bundle.name,
-                      self.mpcenv_bundle.component, self.name)
-
-        # BundleEKFWithController <-> MatcherController
-        component.make_in_port(self.ekf_bundle.name)  # receive state from BundleEKF
-        component.make_out_port(self.ekf_bundle.name)  # send u to BundleEKF
-        self.ekf_bundle.component.make_in_port(self.name)
-
-        brica.connect(self.ekf_bundle.component, "state", 
-                      component, self.ekf_bundle.name)
-        brica.connect(component, self.ekf_bundle.name,
-                      self.ekf_bundle.component, self.name)
-
-        # BundlePlan -> MatcherController
-        component.make_in_port(self.plan_bundle.name)  # receive state from BundlePlan
-        brica.connect(self.plan_bundle.component, "state", 
-                      component, self.plan_bundle.name)
-        
-        self.component = component
-
-    def __call__(self, inputs):
-        self.update(inputs)
-
-        return self.results
-
-    def update(self, inputs):
-        ekf_state = inputs[self.ekf_bundle.name]
-        plan_state = inputs[self.plan_bundle.name]
-        
-        x_plan = plan_state["x"]
-        u_plan = plan_state["u"]
-        K_plan = plan_state["K"]
-        plan_time_stamp = plan_state["time_stamp"]
-        plan_time_id = plan_state["time_id"]
-
-        mu = ekf_state["mu"]
-        ekf_time_stamp = ekf_state["time_stamp"]
-        ekf_time_id = ekf_state["time_id"]
-
-        u = u_plan + K_plan @ (mu - x_plan)
-
-        self.results[self.mpcenv_bundle.name]["u"] = u
-        self.results[self.ekf_bundle.name]["u"] = u
-
-
-class Plan(object):
-    def __init__(self, x_list, u_list, K_list, time_stamp, time_id):
-        self.x_list = x_list
-        self.u_list = u_list
-        self.K_list = K_list
-        self.time_stamp = time_stamp
-        self.time_id = time_id
-
-
-class MatcherILQR(object):
-    def __init__(self, name, ekf_bundle, plan_bundle,
-                 dynamics, cost, dt, T, iter_max):
-        self.name = name
-        self.results = {}
-
-        self.ekf_bundle = ekf_bundle
-        self.plan_bundle = plan_bundle
-        self.bundles = (ekf_bundle, plan_bundle)
-
-        self.results[self.plan_bundle.name] = {}
-
-        self.ilqg = iLQG(dynamics=dynamics, cost=cost, dt=dt)
-        
-        self.T = T
-        self.iter_max = iter_max
-
-        self.last_u_list = None
-        self.last_time_id = 0
-        
-        self.update_component()
-
-    def update_component(self):
-        component = Component(self)
-
-        # BundleEKFWithController -> MatcherILQR
-        component.make_in_port(self.ekf_bundle.name)  # receive state from BundleEKF
-        brica.connect(self.ekf_bundle.component, "state", component,
-                      self.ekf_bundle.name)
-
-        # MatcherController -> BundlePlan
-        component.make_out_port(self.plan_bundle.name)  # send Plan to BundlePlan
-        self.plan_bundle.component.make_in_port(self.name)
-        brica.connect(component, self.plan_bundle.name,
-                      self.plan_bundle.component, self.name)
-        
-        self.component = component
-
-    def __call__(self, inputs):
-        self.update(inputs)
-
-        return self.results
-
-    def update(self, inputs):
-        ekf_state = inputs[self.ekf_bundle.name]
-        
-        mu = ekf_state["mu"]
-        time_stamp = ekf_state["time_stamp"]
-        time_id = ekf_state["time_id"]
-
-        # Initial control sequence for MPC
-        u0 = np.zeros((self.T, 1), dtype=np.float32)
-        if self.last_u_list is not None:
-            # Set initial control signals by copying last control signals.
-            advance_time_id = time_id - self.last_time_id
-            u0[:self.T-advance_time_id,:] = self.last_u_list[advance_time_id:,:]
-
-        x_list, u_list, K_list = self.ilqg.optimize(mu,
-                                                    u0,
-                                                    self.T,
-                                                    self.iter_max)
-        plan = Plan(x_list, u_list, K_list, time_stamp, time_id)
-        self.results[self.plan_bundle.name]["plan"] = plan
-        
-        self.last_u_list = u_list
-        self.last_time_id = time_id
-
-
-class BundlePlan(Bundle):
-    def __init__(self, name, x_dim, u_dim, dt, control_T, plan_src_name):
-        super(BundlePlan, self).__init__(name)
-        self.x_dim = x_dim
-        self.u_dim = u_dim
-        self.dt = dt
-        self.control_T = control_T
-        self.plan_src_name = plan_src_name
-        
-        self.latest_plan = None
-        self.index_in_plan = 0
-        
-        self.update_component()
-
-    def __call__(self, inputs):
-        if self.plan_src_name in inputs.keys() and inputs[self.plan_src_name] is not None:
-            plan = inputs[self.plan_src_name]["plan"]
-            if (self.latest_plan is None) or (self.latest_plan.time_id != plan.time_id):
-                self.latest_plan = plan
-                self.index_in_plan = self.control_T - 1
-                # Note: To make the timestamp equal to that from EKFBundle, we need to start
-                # index_in_plan as 'control_T + 1', but we are now using 'control_T - 1'
-        
-        if self.latest_plan is not None:
-            x = self.latest_plan.x_list[self.index_in_plan]
-            u = self.latest_plan.u_list[self.index_in_plan]
-            K = self.latest_plan.K_list[self.index_in_plan]
-            time_stamp = self.latest_plan.time_stamp + self.dt * self.index_in_plan
-            time_id = self.latest_plan.time_id + self.index_in_plan
-            self.index_in_plan += 1
-        else:
-            x = np.zeros((self.x_dim,), dtype=np.float32)
-            u = np.zeros((self.u_dim,), dtype=np.float32)
-            K = np.zeros((self.u_dim, self.x_dim), dtype=np.float32)
-            time_stamp = 0
-            time_id = 0
-        
-        results = {
-            "x": x,
-            "u": u,
-            "K": K,
-            "time_stamp": time_stamp,
-            "time_id": time_id
-        }
-        return {"state": results}
 
 
 class PendulumObservation(object):
